image/qwen-image-edit/requirements.txt
pydantic >= 2.0.0  # required for inference.sh
inferencesh==0.2.7
torch==2.8.0
git+https://github.com/huggingface/diffusers.git@schedulers/unipc-custom-sigmas
transformers
huggingface_hub>=0.24.0
accelerate>=0.30.0
Pillow>=10.0.0
sentencepiece
protobuf
safetensors
tqdm
einops
gguf>=0.13.0  # For GGUF file handling
requests  # For downloading images from URIs
torchvision